to-hibernate-2

+++ src/main/java/com/me/odontologo/repository/IPacienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface IPacienteRepository extends JpaRepository<Paciente, Long> {
 }
