+++ src/main/java/com/me/odontologo/repository/ITurnoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITurnoRepository extends JpaRepository<Turno, Integer> {
+public interface ITurnoRepository extends JpaRepository<Turno, Long> {
 }
